fix 181 and other discovered issues

+++ modules/core/src/main/scala/codec/NumericCodecs.scala
@@ -19,7 +19,7 @@ trait NumericCodecs {
   val numeric: Codec[BigDecimal] = Codec.simple(_.toString, BigDecimal(_).asRight, Type.numeric)
   def numeric(precision: Int, scale: Int = 0): Codec[BigDecimal] = Codec.simple(_.toString, BigDecimal(_).asRight, Type.numeric(precision, scale))
 
-  val float4: Codec[Double]  = Codec.simple(_.toString, _.toDouble.asRight, Type.float4)
+  val float4: Codec[Float]   = Codec.simple(_.toString, _.toFloat.asRight, Type.float4)
   val float8: Codec[Double]  = Codec.simple(_.toString, _.toDouble.asRight, Type.float8)
 
 }
